// packages/providers/src/pbyp/directus-schema.ts
export const PBYP_SCHEMA: PbypSchema = {
// ...
  air_bookings: {
    primaryKey: "id",
    fields: {
// ...
      shipping_status: {
        type: "string",
        required: true,
        choices: ["CANCELED", "CREATED", "IN_TRANSIT", "ARRIVED_AT_POL"],
      },
// ...
    },
  },
// ...
};